ACPI: Fixing pointer arithmetic
Several adjustements were necessary ;o)

+++ com32/gplinclude/acpi/dsdt.h
@@ -18,7 +18,7 @@
 #define DSDT "DSDT"
 
 typedef struct {
-    uint8_t *address;
+    uint64_t *address;
     s_acpi_description_header header;
     uint8_t *definition_block;
     bool valid;
