+++ com32/gplinclude/acpi/ecdt.h
@@ -21,7 +21,7 @@
 #define EC_ID_OFFSET 65
 
 typedef struct {
-    uint8_t *address;
+    uint64_t *address;
     s_acpi_description_header header;
     bool valid;
     uint32_t warning_energy_level;
