+++ com32/gplinclude/acpi/xsdt.h
@@ -22,7 +22,7 @@ enum { XSDT_TABLE_FOUND };
 typedef struct {
     uint8_t *address;
     s_acpi_description_header header;
-    uint8_t *entry[255];
+    uint64_t *entry[255];
     uint8_t entry_count;
     bool valid;
 } s_xsdt;
